vc++2010fixes
Added a VC++ 2010 solution. The solution has three projects: fhe, helib,
and ntl. The ntl project is not provided, and you must build and add
this static library yourself as it is not a part of helib. Made a few
minor mods in order for the projects to compile properly. fhe is the
executable, and it executes Test_General.

+++ src/FHEContext.h
@@ -140,7 +140,7 @@ public:
 
   // FIXME: run-time error when ithPrime(i) returns 0
   //! @brief Returns the natural logarithm of the ith prime
-  double logOfPrime(unsigned i) const { return log(ithPrime(i)); }
+  double logOfPrime(unsigned i) const { return log((double)ithPrime(i)); }
 
   //! @brief Returns the natural logarithm of productOfPrimes(s)
   double logOfProduct(const IndexSet& s) const {
